Add RequestOptions to Azure.Core
