--- sdk/core/Azure.Core.Experimental/src/RequestOptions.cs
@@ -1,44 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System.Collections.Generic;
-using System.Threading;
-using Azure.Core;
-using Azure.Core.Pipeline;
-
-namespace Azure
-{
-    /// <summary>
-    /// Options which can be used to control the behavior of a request sent by a client.
-    /// </summary>
-    public class RequestOptions
-    {
-        internal List<(HttpPipelinePosition Position, HttpPipelinePolicy Policy)>? Policies { get; private set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RequestOptions"/> class.
-        /// </summary>
-        public RequestOptions()
-        {
-        }
-
-        /// <summary>
-        /// Controls under what conditions the operation raises an exception if the underlying response indicates a failure.
-        /// </summary>
-        public ErrorOptions ErrorOptions { get; set; } = ErrorOptions.Default;
-
-        /// <summary>
-        /// Adds an <see cref="HttpPipelinePolicy"/> into the pipeline for the duration of this request.
-        /// The position of policy in the pipeline is controlled by <paramref name="position"/> parameter.
-        /// If you want the policy to execute once per client request use <see cref="HttpPipelinePosition.PerCall"/>
-        /// otherwise use <see cref="HttpPipelinePosition.PerRetry"/> to run the policy for every retry.
-        /// </summary>
-        /// <param name="policy">The <see cref="HttpPipelinePolicy"/> instance to be added to the pipeline.</param>
-        /// <param name="position">The position of the policy in the pipeline.</param>
-        public void AddPolicy(HttpPipelinePolicy policy, HttpPipelinePosition position)
-        {
-            Policies ??= new();
-            Policies.Add((position, policy));
-        }
-    }
-}
